requirements.txt
rdkit
pandas
tqdm
pillow
stqdm

